Added asio and blake3

+++ src/Blockchain/CTransaction.h
@@ -13,11 +13,10 @@
 #include "Crypto/CCryptoUtils.h"
 
 namespace DeFile::Blockchain {
-    //By the version 1 transaction definition, a raw, non-signed, non-serialized transaction is 173 bytes long.
-    //The final size of a fully signed, serialized and stringified transaction is 552 bytes, and it's stored in the final block.
+    //By the version 1 transaction definition, a raw, non-signed, non-serialized transaction is 176 bytes long.
     class CTransaction {
         private:
-            uint8_t mVersion;
+            uint32_t mVersion;
             std::string mSourceAddress;
             std::string mDestinationAddress;
             uint64_t mTransferredAmount;
@@ -29,7 +28,7 @@ namespace DeFile::Blockchain {
             
             uint16_t mTxSize; // Size of the transaction. This should only be accessed after hashing.
         public:
-            CTransaction(uint8_t version, const std::string &srcAddr, const std::string &destAddr, uint64_t amount, uint64_t srcBal, uint64_t destBal)
+            CTransaction(uint32_t version, const std::string &srcAddr, const std::string &destAddr, uint64_t amount, uint64_t srcBal, uint64_t destBal)
              : mVersion(version), mSourceAddress(srcAddr), mDestinationAddress(destAddr), mTransferredAmount(amount), mSourceNewBalance(srcBal), mDestinationNewBalance(destBal), mTimestamp(CTimeUtils::getUnixTimestampNS()) {
                 memset(mTxHash, 0, SHA256_DIGEST_LENGTH);     // mHash nulls 
             }
@@ -51,7 +50,7 @@ namespace DeFile::Blockchain {
                 return ss.str();
             }
 
-            uint8_t getVersion() { return mVersion; }
+            uint32_t getVersion() { return mVersion; }
             std::string getSourceAddress() { return mSourceAddress; }
             std::string getDestinationAddress() { return mDestinationAddress; }
             uint64_t getTransferedAmount() { return mTransferredAmount; }
